Kenzie/Sub-MAF-07: Updated logic for color theme and got rid of other classes for now.
