+++ app/build.gradle
@@ -65,4 +65,5 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     implementation "androidx.preference:preference-ktx:1.1.1"
+    implementation "com.google.android.material:material:1.11.0"
 }
