+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:icon="@mipmap/blueharvest"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Light"
+        android:theme="@style/DynamicColors.Overlay_Light"
     tools:targetApi="31"
         android:enableOnBackInvokedCallback="true"
         android:localeConfig="@xml/locales_config">
